capture/requirements.txt
# Skylapse Capture Service Dependencies

# Core async framework
asyncio-mqtt>=0.11.1

# Web framework for API server
aiohttp>=3.8.0

# Configuration management
PyYAML>=6.0

# Image processing (basic support)
Pillow>=9.0.0

# System integration
psutil>=5.9.0

# Logging and monitoring
structlog>=22.3.0

# Future camera support (commented out for Sprint 1)
# picamera2>=0.3.0  # Pi camera support
# libcamera>=0.0.1  # Direct camera access

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Future enhancements (Phase 2)
# requests>=2.28.0  # Weather API integration
# numpy>=1.21.0     # Image analysis
# opencv-python>=4.6.0  # Advanced image processing